backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
anthropic>=0.35.0claude-agent-sdk>=0.1.0
